+++ master/protocol/replay_protocol.cc
@@ -1,6 +1,6 @@
 #include "protocol.hh"
 
-char *MasterProtocol::replaySet( size_t &size, uint16_t instanceId, uint32_t requestId, char *key, uint8_t keySize, char *value, uint8_t valueSize ) {
+char *MasterProtocol::replaySet( size_t &size, uint16_t instanceId, uint32_t requestId, char *key, uint8_t keySize, char *value, uint32_t valueSize ) {
 	size = this->generateKeyValueHeader(
 		PROTO_MAGIC_REQUEST,
 		PROTO_MAGIC_TO_MASTER,
@@ -28,7 +28,7 @@ char *MasterProtocol::replayGet( size_t &size, uint16_t instanceId, uint32_t req
 	return this->buffer.recv;
 }
 
-char *MasterProtocol::replayUpdate( size_t &size, uint16_t instanceId, uint32_t requestId, char *key, uint8_t keySize, char *valueUpdate, uint8_t valueUpdateOffset, uint32_t valueUpdateSize) {
+char *MasterProtocol::replayUpdate( size_t &size, uint16_t instanceId, uint32_t requestId, char *key, uint8_t keySize, char *valueUpdate, uint32_t valueUpdateOffset, uint32_t valueUpdateSize) {
 	size = this->generateKeyValueUpdateHeader(
 		PROTO_MAGIC_REQUEST,
 		PROTO_MAGIC_TO_MASTER,
